Moved wcag guid constant to wcag-taxonomy-provider
