+++ src/axe-tool-property-provider-21.test.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { getAxeToolProperties21 } from './axe-tool-property-provider-21';
-import { WcagGuid } from './constants';
 import * as Sarif from './sarif/sarif-2.1.2';
+import { WcagGuid } from './wcag-taxonomy-provider';
 
 describe('axe-tool-property-provider 2.1', () => {
     describe('getAxeToolProperties21', () => {
