--- src/constants.ts
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-export const WcagGuid = 'ca34e0e1-5faf-4f55-a989-cdae42a98f18';
