+++ src/wcag-taxonomy-provider.ts
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { WcagGuid } from './constants';
 import { DictionaryStringTo } from './dictionary-types';
 import * as Sarif from './sarif/sarif-2.1.2';
 import { WCAGLinkData } from './wcag-link-data';
 
+export const WcagGuid = 'ca34e0e1-5faf-4f55-a989-cdae42a98f18';
 export function getWcagTaxonomy(
     sortedWcagTags: string[],
     tagsToWcagLinkData: DictionaryStringTo<WCAGLinkData>,
